graderng/requirements.txt
celery==4.3.0
django-cors-headers==3.2.0
django-filebrowser-no-grappelli==3.7.8
Django==2.2.6
djangorestframework==3.10.3
djongo==1.2.36
fakeredis==1.2.1
gunicorn==20.0.0
lupa==1.9
redis==3.3.11
requests==2.22.0
sqlparse==0.2.4
whitenoise==4.1.4
